Fixes issue with the lane connectivity size being messed up because it wasn't taking the edge elevation samples into account

+++ valhalla/baldr/graphtile.h
@@ -632,7 +632,7 @@ public:
    * @return String containing encoded elevation samples. Empty if there are no elevation samples.
    * @throws std::range_error if index is out-of-bounds
    */
-  std::string_view elevationSampleAtIndex(uint16_t directedEdgeIndex) const;
+  std::string_view elevationSampleAtIndex(uint32_t directedEdgeIndex) const;
 
   /**
    * Convenience method to get the offset into the text table for the turn lanes
